update: 更新文件 ng_shader_color_gradient_effect_test.cpp

+++ graphic_test/test/rs_effect_feature/ng_shader_test/ng_shader_color_gradient_effect_test.cpp
@@ -170,7 +170,7 @@ GRAPHIC_TEST(NGShaderColorGradientEffectTest, EFFECT_TEST, Set_Color_Gradient_Ef
         colorGradient->Setter<ColorGradientEffectColor2Tag>(Vector4f{0.0f, 0.0f, 1.0f, 1.0f});
         colorGradient->Setter<ColorGradientEffectColor3Tag>(Vector4f{1.0f, 1.0f, 0.0f, 1.0f});
 
-        SetEffectChildNode(i, columnCount, rowCount, effectNode, colorGradient);
+        SetEffectChildNode(static_cast<size_t>(i), columnCount, rowCount, effectNode, colorGradient);
     }
 }
 
@@ -194,7 +194,7 @@ GRAPHIC_TEST(NGShaderColorGradientEffectTest, EFFECT_TEST, Set_Color_Gradient_Ef
         colorGradient->Setter<ColorGradientEffectColor2Tag>(Vector4f{0.0f, 0.0f, 1.0f, 1.0f});
         colorGradient->Setter<ColorGradientEffectColor3Tag>(Vector4f{1.0f, 1.0f, 0.0f, 1.0f});
 
-        SetEffectChildNode(i, columnCount, rowCount, effectNode, colorGradient);
+        SetEffectChildNode(static_cast<size_t>(i), columnCount, rowCount, effectNode, colorGradient);
     }
 }
 
@@ -213,7 +213,7 @@ GRAPHIC_TEST(NGShaderColorGradientEffectTest, EFFECT_TEST, Set_Color_Gradient_Ef
         colorGradient->Setter<ColorGradientEffectBrightnessTag>(extremeValues[i]);
         colorGradient->Setter<ColorGradientEffectColorNumberTag>(4.0f);
         
-        SetEffectChildNode(i, columnCount, rowCount, effectNode, colorGradient);
+        SetEffectChildNode(static_cast<size_t>(i), columnCount, rowCount, effectNode, colorGradient);
     }
 }
